+++ deep_quoridor/src/agents/__init__.py
@@ -6,7 +6,7 @@ __all__ = [
     "DExpAgent",
     "FlatDQNAgent",
     "HumanAgent",
-    "NewDexpAgent",
+    "NDexpAgent",
     "RandomAgent",
     "ReplayAgent",
     "ReplayBuffer",
@@ -27,7 +27,7 @@ from agents.dexp import DExpAgent  # noqa: E402
 from agents.flat_dqn import FlatDQNAgent  # noqa: E402
 from agents.greedy import GreedyAgent  # noqa: E402, F401
 from agents.human import HumanAgent  # noqa: E402, F401
-from agents.pagent import NewDexpAgent
+from agents.ndexp import NDexpAgent
 from agents.random import RandomAgent  # noqa: E402, F401
 from agents.replay import ReplayAgent  # noqa: E402, F401
 from agents.sb3_ppo import SB3PPOAgent  # noqa: E402, F401
@@ -37,7 +37,7 @@ AgentRegistry.register("dexp", DExpAgent)
 AgentRegistry.register("flatdqn", FlatDQNAgent)
 AgentRegistry.register("greedy", GreedyAgent)
 AgentRegistry.register("human", HumanAgent)
+AgentRegistry.register("ndexp", NDexpAgent)
 AgentRegistry.register("random", RandomAgent)
 AgentRegistry.register("simple", SimpleAgent)
 AgentRegistry.register("sb3ppo", SB3PPOAgent)
-AgentRegistry.register("newdexp", NewDexpAgent)
